# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TFSO API",
        "description": "API for TFSO application",
        "contact": {
            "name": "Burak",
            "url": "https://burakkaynar.com",
            "email": "burak@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/djs": {
            "get": {
                "tags": [
                    "DJs"
                ],
                "summary": "Get all DJs",
                "operationId": "5dc880d05a9254aa104d61a57a549747",
                "responses": {
                    "200": {
                        "description": "A list of DJs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "DJs"
                ],
                "summary": "Create a new DJ",
                "operationId": "3da087d8b355291ba7bef0217dbc0ba5",
                "requestBody": {
                    "description": "Data for the new DJ",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Yeni DJ"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "yeni.dj@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "profile_photo": {
                                        "type": "string",
                                        "example": "path/to/photo.jpg"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Yeni DJ'in biyografisi."
                                    },
                                    "instagram": {
                                        "type": "string",
                                        "example": "yenidj"
                                    },
                                    "twitter": {
                                        "type": "string",
                                        "example": "yenidj"
                                    },
                                    "facebook": {
                                        "type": "string",
                                        "example": "https://facebook.com/yenidj"
                                    },
                                    "tiktok": {
                                        "type": "string",
                                        "example": "yenidj"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "DJ created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "Model representing a user in the system",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the user",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true
                    },
                    "name": {
                        "description": "Name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email address of the user",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Password for the user account",
                        "type": "string",
                        "format": "password"
                    },
                    "profile_photo": {
                        "description": "URL of the user's profile photo",
                        "type": "string"
                    },
                    "bio": {
                        "description": "Short biography of the user",
                        "type": "string"
                    },
                    "instagram": {
                        "description": "Instagram handle of the user",
                        "type": "string"
                    },
                    "twitter": {
                        "description": "Twitter handle of the user",
                        "type": "string"
                    },
                    "facebook": {
                        "description": "Facebook profile URL of the user",
                        "type": "string"
                    },
                    "tiktok": {
                        "description": "TikTok handle of the user",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}